added sd card logging logs meaningless data when button is pressed

+++ Car/src/SensorData.h
@@ -5,10 +5,9 @@ enum sensorIDs { INERTIA_SWITCH, KEY_SWITCH, ACCEL_PEDAL, BRAKE_PEDAL };
 
 class SensorData {
     public:
-        
         enum sensorIDs sensorid;
         double data; //not sure about this data type
-        double timestamp;
+        unsigned long timestamp;
 };
